ci: Add workflow to manually publish docker image
You can now manually publish a docker image to ECR by going to the
Actions tab and selecting the "Manually Publish Aviatrix AWS HA Docker
Image" workflow. This will build and publish the docker image to ECR
with the following tags: `build-<git sha>` and `<ref tag>`.

The ref tag is the branch that was used when the workflow was triggered.
The git sha is the sha of the commit on that branch.